+++ components/organisms/UserSettingsPage/user-settings-page.tsx
@@ -10,13 +10,14 @@ import Select from "components/atoms/Select/select";
 import SelectOption from "components/atoms/Select/select-option";
 import LanguagePill from "components/atoms/LanguagePill/LanguagePill";
 
-import { updateUser, updateUserPayload } from "lib/hooks/update-user";
+import { updateUser, UpdateUserPayload } from "lib/hooks/update-user";
 import { ToastTrigger } from "lib/utils/toast-trigger";
 import { authSession } from "lib/hooks/authSession";
 import { validateEmail } from "lib/utils/validate-email";
 import { timezones } from "lib/utils/timezones";
 import { updateEmailPreferences } from "lib/hooks/updateEmailPreference";
 import { useFetchUser } from "lib/hooks/useFetchUser";
+import getInterestOptions from "lib/utils/getInterestOptions";
 
 interface userSettingsPageProps {
   user: User | null;
@@ -42,7 +43,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
   });
   const [selectedInterest, setSelectedInterest] = useState<string[]>([]);
   const formRef = useRef<HTMLFormElement>(null);
-  const interestArray = ["javascript", "python", "rust", "ML", "AI", "react"];
+  const interestArray = getInterestOptions();
 
   useEffect(() => {
     async function fetchAuthSession() {
@@ -56,6 +57,8 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
         formRef.current!.twitter_username.value = response.twitter_username;
         formRef.current!.company.value = response.company;
         formRef.current!.location.value = response.location;
+        formRef.current!.github_sponsors_url.value = response.github_sponsors_url;
+        formRef.current!.linkedin_url.value = response.linkedin_url;
       }
     }
     fetchAuthSession();
@@ -116,7 +119,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
 
   const handleUpdateProfile = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    const payload: updateUserPayload = {
+    const payload: UpdateUserPayload = {
       name: formRef.current!.nameInput.value,
       email,
       bio: formRef.current!.bio.value,
@@ -126,7 +129,11 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
       location: formRef.current!.location.value,
       // eslint-disable-next-line camelcase
       display_local_time: displayLocalTime,
-      timezone
+      timezone,
+      // eslint-disable-next-line camelcase
+      github_sponsors_url: formRef.current!.github_sponsors_url.value !== "" ? formRef.current!.github_sponsors_url.value : undefined,
+      // eslint-disable-next-line camelcase
+      linkedin_url: formRef.current!.linkedin_url.value !== "" ? formRef.current!.linkedin_url.value: undefined
     };
     if (formRef.current?.url.value) {
       payload.url = formRef.current!.url.value;
@@ -185,8 +192,22 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
               name="url"
             />
             <TextInput
+              classNames="bg-light-slate-4 text-light-slate-11 font-medium"
+              placeholder="https://github.com/sponsors/open-sauced"
+              label="GitHub Sponsors URL"
+              pattern="http[s]?://.*\..{2,}"
+              name="github_sponsors_url"
+            />
+            <TextInput
+              classNames="bg-light-slate-4 text-light-slate-11 font-medium"
+              placeholder="https://www.linkedin.com/in/brianldouglas"
+              label="LinkedIn URL"
+              pattern="http[s]?://.*\..{2,}"
+              name="linkedin_url"
+            />                        
+            <TextInput
               classNames="bg-light-slate-4 text-light-slate-11"
-              placeholder="@saucedopen"
+              placeholder="saucedopen"
               label="Twitter Username"
               name="twitter_username"
             />